Implementation of k8s-discovery datagatherer
This makes use of the discovery client in client-go to get the server
version information. We need this to use in upcoming packages.

https://godoc.org/k8s.io/client-go/discovery

Sadly, the fake discovery client doesn't seem to work in the same way as
the dynamic one and I've not spent the time to write tests around it.
However, it's operation is much simpler. Open to suggestions on how to
test this in the same way as the dynamic client.

This PR also updates all docs and renames the k8s datagatherer to
k8s-dynamic in a backwards compatible manner.

(Internal Issue: https://github.com/jetstack/preflight-platform/issues/350)

Signed-off-by: Charlie Egan <charlieegan3@users.noreply.github.com>