Restructure gatherAndPostData to allow local write
gatherAndPostData has been split into:
gatherData which returns config information, the prefight client
and readings.

postData which sends this data to preflight server as previously
happened in gatherAndPostData

The outputpath flag now determines if postData is called.
If the flag is set, postData is not called and the data is written
locally.

This also fixes the issue of the flag only working when no org is set.

Signed-off-by: Jamie Leppard <JammyL@users.noreply.github.com>