Adds write-to-file flag to store data locally (#180)
* Adds write-to-file flag to store data locally

Only stores data locally if no organization is specified.
This can be added if needed.

Signed-off-by: Jamie Leppard <JammyL@users.noreply.github.com>

* Restructure gatherAndPostData to allow local write

gatherAndPostData has been split into:
gatherData which returns config information, the prefight client
and readings.

postData which sends this data to preflight server as previously
happened in gatherAndPostData

The outputpath flag now determines if postData is called.
If the flag is set, postData is not called and the data is written
locally.

This also fixes the issue of the flag only working when no org is set.

Signed-off-by: Jamie Leppard <JammyL@users.noreply.github.com>

* Breaks down gatherData into smaller functions

gatherData is broken down into:

getConfiguration - returns config and preflight client
this is now implemented outside of gatherAndOutputData
in the main Run loop

gatherData - creates data gatherers and generates readings
This is run inside gatherAndOutputData as previously done

Signed-off-by: Jamie Leppard <JammyL@users.noreply.github.com>

Co-authored-by: Jamie Leppard <JammyL@users.noreply.github.com>