Run datagather informers 'in background'
This updates the datagatherers to be started and run in a non-blocking
manner.

We allow datagatherers 5 seconds to complete an initial sync or to
error. Once this has passed, we proceed to the initial gather and
upload. We acknowledge that this might mean incomplete data is sent to
the backend, however this acceptable since it could happen for other
reasons in any of our previous architectures.

Signed-off-by: Charlie Egan <charlieegan3@users.noreply.github.com>