New CI workflow
* Moves compilation of binaries into hack/build.sh. In this file, new architectures can be compiled by adding a new
function to the script. There are instructions in comments
* Removes the docker-in-docker build. Instead, the CI now installs go 1.13.7 and compiles the binaries directly
on the actions host
* Merges the release-master and release-tag workflows. Instead, a single workflow is triggered that will push to
the git SHA or to master. If a tag has been pushed, it should use the tag instead.
* Removes the docker_build job from the test workflow. Using go test ./... should pick up on any go build failures

Personally, I find this a much clearer way to manage the compilation and docker build process. Instead of a large
Makefile with lots of templating, we have a single bash script for handling cross-compilation that can be added
to easily. Likewise, the docker build process also has a single bash script.

Signed-off-by: David Bond <davidsbond93@gmail.com>