Use USER 1000:1000 in agent image.
We're running issues getting the agent to work out of the box on OpenShift clusters due to the
way they set up security contexts. We're trying to align our Docker image with cert-manager's
as their deployments work out of the box with the restrictions that OpenShifts's SecurityContextConstraints
custom resource applies by default.

Signed-off-by: David Bond <davidsbond93@gmail.com>