Modify CI to push multiple tags
The retagging commands don't work with quay.io, so instead we can use the multiple --tag parameters allowed
in the `docker buildx build` command. This makes the `release-tag` workflow redundant, as all it did was
perform that retagging.

Now, when a commit is merged to master, an image should be created for `latest`, `canary` and the commit SHA. When
a tag is pushed, an image should be created for `latest`, `canary` and the tag.

Signed-off-by: David Bond <davidsbond93@gmail.com>