Add signing, provenance and SBOM
Signs the agent container image with cosign's keyless signing and Github
OIDC. Also signs and attaches a SLSA provenance attestation and a
CycloneDX SBOM.

The provenance provided takes the agent up to SLSA 1.

The signatures and other attachments are stored in a separate repository
to the actual container images. There are two main reasons for this:
1. Although quay.io supports cosign signatures as of the latest release,
   it doesn't support uploading attestations or SBOMs.
2. Pulling SBOMs, attestations and signatures for verification can skew image
   pull metrics for the repository. The number of pulls strikes me as
   something we might want to track reliably for the agent.

Signed-off-by: Rob Best <robertbest89@gmail.com>