feat: Adding licensing information to image
Using go-license to pull in the licensing for all go dependencies. The
final image then hosts all licencing informations under
"/usr/share/doc/preflight".

For more minimal impact on the repo this was added to the builder image.
This then gets output into "./builds/licenses" which is already copied
out by the makefile. Adds a step to copy over the licencing in the final
release image.