Remove replace directives and get modules upgraded
This commit removes the replace directives within the go.mod file and upgrades
dependencies that can be upgraded. It also upgrades the module to go 1.18 and
removes deprecated tools like golint.

The rationale here is that we've left this module for so long things have gotten
quite out of date and automatic upgrades now usually fail due to dependencies
pinned by the replace directive. We also depend on this module internally
so the issues propagate there too.

Signed-off-by: David Bond <davidsbond93@gmail.com>