Remove all unused data gatherers
This commit removes all datagatherers except for the kubernetes dynamic
and discovery gatherers.

As the project has changed over time we've moved away from production readiness
checks, which the removed datagatherers were primarily used for. A default
agent configuration will not include any of these, so we can start to remove
them.

Signed-off-by: David Bond <davidsbond93@gmail.com>