fix: runAsUser in yaml is unneeded as its set in the Dockerfile
This also fixes an issue in OpenShift where OpenShift wants you to use a certain range of UIDs that 1000 does not fall into. By not specifying an runAsUser in the yaml, OpenShift picks a appropriate UID for you.

Since the Dockerfile contains the USER directive this should have no effect on non OpenShift Kubernetes clusters.

Signed-off-by: Adam Talbot <adam.talbot@venafi.com>