makefile-modules: migrate the Makefile to makefile-modules
Initially, my goal was to just renamed the Makefile to make/02_mod.mk,
and change nothing else, with the intent of migrating bit by bit. After
a few attempts, I found that the fact that the Makefile is being run
within a container makes things needlessly complex, and trying to make
makefile-modules work in that context isn't worth it. That's why I
propose to migrate everything at once, with the goal of making no
breaking changes to the Helm charts and containers (except for the
binary location, binary name, entrypoint, and cmd).