disable the dump of the config file on startup
The config is no longer dumped at startup. I don't think there is a
point in dumping the whole content of a file... it's super verbose and
is even hurtful as it drowns actual important errors within a flow of
log lines.

If we think that dumping the config is an important feature, it should
be gated behind a --debug flag. I propose to remove it for now.