disable the `config.server` field when using --venafi-connection
For context, EU customers need to change the URL of the VCP API to point
to https://api.venafi.eu. To do that, and imagining that they are using
the VenafiConnection authentication, they may try to use the
`spec.vcp.url` field on their VenafiConnection resource and find that
this field doesn’t do anything because the Helm chart's `config.server`
is set to https://api.venafi.cloud by default.

Another possible scenario is that EU customers may end up with a
VenafiConnection configured with the `spec.vcp.url` field set to
`https://api.venafi.eu`. This VenafiConnection will have been already
working well with venafi-enhanced-issuer and approver-policy-enterprise.
Once they try to switch the Agent to the VenafiConnection auth method,
they will see that it doesn’t work because the Agent picks up the
default value in the Agent’s helm chart, i.e.,

```
config:
  server: https://api.venafi.cloud.
```