e2e/test.sh: removed the need for VEN_OWNING_TEAM
I spent 3 hours realizing that the reason the script was silently
failing was because I was using an ID rather than a name as
VEN_OWNING_TEAM.

Since the team doesn't really matter in this test, I propose to just
pick the first team that is returned by /v1/teams. That should avoid
having to find out the name or ID of a team to run this script.