feat: add cluster uid derived from kube-system ns
This seems like a simple enough way of uniquely identifying a cluster.
Google searches show that this pattern is used elsewhere.

It seems fair to assume that if the kube-system ns UID changes, it
implies that the cluster changed enough to be considered different
anyway!

This is added to the context so it can be used elsewhere, on the
assumption that it can be determined at startup and won't change
later. We'll use this for "machinehub mode" soon.

Signed-off-by: Ashley Davis <ashley.davis@cyberark.com>