fix: add credentials for pulling private dependency with govulncheck
It seems difficult to stop govulncheck attempting to pull and analyse
private dependencies. We still want to run it, so the simplest thing to
do is diverge from upstream makefile-modules and maintain the
govulncheck workflow by hand in this repo.

This requires changes to the govulncheck workflow itself, and means we
have to disable the upstream govulncheck targets and copy them locally.

Signed-off-by: Ashley Davis <ashley.davis@cyberark.com>